Label Pull Requests (#33)
# Pull Request

## Description

This pull request includes changes to automate the labeling of pull
requests based on modified files and to add size labels to pull requests
based on the number of lines changed. The most important changes include
the addition of a configuration file for the labeler and a workflow file
to handle the pull request tasks.

Labeling and size automation:

*
[`.github/other-configurations/labeller.yml`](diffhunk://#diff-b490f987d9b055ce6e06e37d13763d61f6e0d5ee349b390bbb910f8d88ff36b9R1-R40):
Added configuration for auto-assigning labels to pull requests based on
changed files. This includes labels for documentation, Rust files,
markdown files, dependencies, Justfiles, shell scripts, GitHub Actions,
and git hooks.
*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR1-R36):
Added a workflow to label pull requests and add size labels. This
workflow triggers on pull request events and uses the `actions/labeler`
and `pascalgn/size-label-action` actions to apply the appropriate
labels.

Fixes #22