Add Dependency Review (#34)
# Pull Request

## Description

This pull request introduces a new job to the GitHub Actions workflow to
perform a dependency review. The most important change is the addition
of the `dependency-review` job in the
`.github/workflows/pull-request-tasks.yml` file.

Key changes:

*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR37-R48):
Added a `dependency-review` job to the workflow, which includes steps
for checking out the repository and performing a dependency review using
the `actions/dependency-review-action` at version `v4.5.0`.

Fixes #25