Add Superlinter Scanning (#39)
# Pull Request

## Description

This pull request includes several changes to the configuration files
and workflows to improve code quality checks and linting processes. The
most important changes include the addition of new configuration files
for various linters and the update to the GitHub Actions workflow to
integrate these linters.

### Linter Configuration Additions:

*
[`.github/super-linter-configurations/.ecrc`](diffhunk://#diff-82be02a68d11ecce2e6b048fe6a0c5e2ff985c76196d78c22da7f7c0c8719509R1-R10):
Added a new configuration file for EditorConfig with default settings.
*
[`.github/super-linter-configurations/.jscpd.json`](diffhunk://#diff-cd81db8202b69736e9bf6c719f609c91d4998d9aff32d1e9940adf73db3f29ecR1-R10):
Added a new configuration file for JSCPD (JavaScript Copy/Paste
Detector) with specific settings for minimum and maximum lines, size,
and tokens.
*
[`.github/super-linter-configurations/.yaml-lint.yml`](diffhunk://#diff-d002df7aeed074c70587c0a7a68935b9932e9cda254f58d83a2accea6103b233R1-R8):
Added a new configuration file for YAML linting with rules to disable
certain checks and set minimum spaces for comments.

### Workflow Updates:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R14-R38):
Updated the workflow to include a new job named `check-code-quality`
which runs the Super-Linter on the codebase with the newly added
configurations.

Fixes #16