Update Cargo Dependencies (#42)
# Pull Request

## Description

This pull request includes changes to the Dependabot configuration file
to add support for the Cargo package ecosystem. The most important
change is the addition of a new update configuration for Cargo packages.

Updates to Dependabot configuration:

*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28R20-R36):
Added a new configuration for the Cargo package ecosystem, specifying
the update schedule, target branch, and update types.

Fixes #41