Add Justfile Format Checking (#46)
# Pull Request

## Description

This pull request introduces a new job to the GitHub Actions workflow to
ensure the format of the `Justfile` is checked. The most important
change is the addition of the `check-justfile-format` job in the
`.github/workflows/code-checks.yml` file.

Improvements to the GitHub Actions workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R57-R70):
Added a new job named `check-justfile-format` to verify the format of
the `Justfile`. This job includes steps for checking out the repository,
setting up Just, and running the format check command.

Fixes #19