Add Basic Readme (#60)
# Pull Request

## Description

This pull request includes a small addition to the `README.md` file. The
change introduces a new project called `pls` (Project Listing System)
and provides a brief description of its purpose and functionality.

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R1-R5):
Added a description for the `pls` project, which is a command line tool
for managing projects.

Fixes #8